ai-think-tank/backend/requirements.txt
# Backend requirements
fastapi
uvicorn
pydantic
langgraph
langchain-openai
langchain-core
python-dotenv
sse-starlette
